FEAT: modify according to reviews
